% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_helpers.R
\name{export_data_to_csv}
\alias{export_data_to_csv}
\title{Export data to CSV for visualization}
\usage{
export_data_to_csv(data, file)
}
\arguments{
\item{data}{A data frame prepared for visualization}

\item{file}{Path to save the CSV file}
}
\value{
The path to the saved CSV file
}
\description{
This function exports data to a CSV file for visualization
}
\keyword{internal}
