% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_handlers.R
\name{get_selection}
\alias{get_selection}
\title{Get the currently selected points from the visualization}
\usage{
get_selection()
}
\value{
A data frame containing the selected points
}
\description{
Get the currently selected points from the visualization
}
