% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_server.R
\name{open_viz}
\alias{open_viz}
\title{Open the visualization in RStudio's Viewer pane or browser}
\usage{
open_viz(url = "http://localhost:3000")
}
\arguments{
\item{url}{URL of the visualization, defaults to localhost:3000}
}
\value{
NULL (invisible)
}
\description{
Open the visualization in RStudio's Viewer pane or browser
}
