% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_server.R
\name{run_server}
\alias{run_server}
\title{Start a local Node.js server for development}
\usage{
run_server(port = 3000, open = TRUE)
}
\arguments{
\item{port}{Port number to use for the server}

\item{open}{Whether to open the visualization in browser}
}
\value{
The process ID of the server
}
\description{
This function starts a local Node.js server for development purposes
}
