% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_handlers.R
\name{save_selection}
\alias{save_selection}
\title{Save the current selection to a variable}
\usage{
save_selection(name, envir = parent.frame())
}
\arguments{
\item{name}{Name to assign to the selection}

\item{envir}{Environment to assign the selection to, defaults to parent frame}
}
\value{
Invisibly returns the selection
}
\description{
This function saves the current selection to a variable in the specified environment
}
