% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_handlers.R
\name{store_selection}
\alias{store_selection}
\title{Store selection in package environment}
\usage{
store_selection(selection)
}
\arguments{
\item{selection}{Selection data}
}
\description{
Store selection in package environment
}
\keyword{internal}
